Refactor popup into reusable component
